more refactoring

+++ iteration2/src/course/FacultyTechnicalElectiveCourse.java
@@ -43,7 +43,7 @@ public class FacultyTechnicalElectiveCourse extends ElectiveCourse{
         int noOfCoursesTillSemester = 0;
 
         for(int i = 0; i <= semester; i++){
-            noOfCoursesTillSemester += numberOfCoursesTakeableBySemester.get(i);
+            noOfCoursesTillSemester += numberOfCoursesTakeableBySemester.get(i).intValue();
         }
 
         return noOfCoursesTillSemester;
