+++ iteration2/src/course/TechnicalElectiveCourse.java
@@ -54,7 +54,7 @@ public class TechnicalElectiveCourse extends ElectiveCourse{
         int noOfCoursesTillSemester = 0;
 
         for(int i = 0; i <= semester; i++){
-            noOfCoursesTillSemester += numberOfCoursesTakeableBySemester.get(i);
+            noOfCoursesTillSemester += numberOfCoursesTakeableBySemester.get(i).intValue();
         }
 
         return noOfCoursesTillSemester;
